+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 ## ChangeLog
 
+### 1.5.0
+
+- [bug] Fix #144: `onCheck` and `onUncheck` events are reversed when using `clickToSelect` option. (jQuery 1.7.2 bug).
+- [bug] Fix IE browser display header bug when use `mergeCells` method.
+- [bug] Fix #269: array as row bug.
+- [bug] Fix #314: `rowStyle` bug.
+- [enh] Add de-DE, hu-HU, sk-SK locales.
+- [enh] Fix #261: add namespace to `.table` style.
+- [bug] Fix #160, #323: operate events don't work in card view.
+- [enh] Add `filterBy`, `scrollTo`, `prevPage` and `nextPage`, `check` and `uncheck` methods.
+- [enh] Add `onPreBody` and `onPostBody` events.
+- [enh] Add `searchable` column option.
+- [enh] Fix #59: support load multiple locale files.
+- [enh] Modify the scope of the column events.
+- [enh] Improve editable extension.
+
 ### 1.4.0
 
 - [enh] Fix #119, #123: Save all `id` and `class` of `tr` and `td` for html table.